+++ resources/js/pages/Devices/Create.tsx
@@ -1,0 +1,5 @@
+import AppLayout from '../../layouts/AppLayout';
+import DeviceForm from './DeviceForm';
+
+export default function DevicesCreate() { return <DeviceForm />; }
+DevicesCreate.layout = (page: React.ReactNode) => <AppLayout>{page}</AppLayout>;